requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
elasticsearch
python-dotenv
requests
